81110 pass likers from parent component to LikeButton
